Rearrange tests, add ∇_cent_to_cent tests
