+++ test/runtests.jl
@@ -7,6 +7,7 @@ include("spaces.jl")
 include("field.jl")
 include("fielddiffeq.jl")
 include("operators.jl")
+include("column_operators.jl")
 
 if "CUDA" in ARGS
     include("gpu/cuda.jl")
